Move geetS3EndpointForRegion to utils, change to regex.test instead of string.match for authorizer ARNs
