+++ lib/plugins/aws/deploy/lib/validateTemplate.js
@@ -1,5 +1,5 @@
 'use strict';
-const getS3EndpointForRegion = require('./getS3EndpointForRegion');
+const getS3EndpointForRegion = require('../../utils/getS3EndpointForRegion');
 
 module.exports = {
   validateTemplate() {
