+++ lib/plugins/aws/lib/updateStack.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const BbPromise = require('bluebird');
-const getS3EndpointForRegion = require('../deploy/lib/getS3EndpointForRegion');
+const getS3EndpointForRegion = require('../utils/getS3EndpointForRegion');
 
 const NO_UPDATE_MESSAGE = 'No updates are to be performed.';
