+++ lib/plugins/aws/package/compile/events/apiGateway/lib/authorizers.js
@@ -25,7 +25,7 @@ module.exports = {
         const authorizerLogicalId = this.provider.naming.getAuthorizerLogicalId(authorizer.name);
 
         if (typeof authorizer.arn === 'string'
-          && authorizer.arn.match(awsArnRegExs.cognitoIdpArnExpr)) {
+          && awsArnRegExs.cognitoIdpArnExpr.test(authorizer.arn)) {
           authorizerProperties.Type = 'COGNITO_USER_POOLS';
           authorizerProperties.ProviderARNs = [authorizer.arn];
         } else {
