+++ lib/plugins/aws/package/compile/events/apiGateway/lib/method/authorization.js
@@ -20,7 +20,7 @@ module.exports = {
       let authorizationType;
       const authorizerArn = http.authorizer.arn;
       if (typeof authorizerArn === 'string'
-        && authorizerArn.match(awsArnRegExs.cognitoIdpArnExpr)) {
+        && awsArnRegExs.cognitoIdpArnExpr.test(authorizerArn)) {
         authorizationType = 'COGNITO_USER_POOLS';
       } else {
         authorizationType = 'CUSTOM';
