+++ lib/plugins/aws/package/compile/events/apiGateway/lib/permissions.js
@@ -46,7 +46,7 @@ module.exports = {
           .getLambdaApiGatewayPermissionLogicalId(authorizer.name);
 
         if (typeof authorizer.arn === 'string'
-          && authorizer.arn.match(awsArnRegExs.cognitoIdpArnExpr)) {
+          && awsArnRegExs.cognitoIdpArnExpr.test(authorizer.arn)) {
           return;
         }
