+++ lib/plugins/aws/package/compile/events/apiGateway/lib/validate.js
@@ -256,7 +256,7 @@ module.exports = {
     const integration = this.getIntegration(http);
     if (integration === 'AWS_PROXY'
       && typeof arn === 'string'
-      && arn.match(awsArnRegExs.cognitoIdpArnExpr)
+      && awsArnRegExs.cognitoIdpArnExpr.test(arn)
       && authorizer.claims) {
       const errorMessage = [
         'Cognito claims can only be filtered when using the lambda integration type',
