+++ lib/plugins/aws/provider/awsProvider.js
@@ -12,7 +12,7 @@ const https = require('https');
 const fs = require('fs');
 const objectHash = require('object-hash');
 const PromiseQueue = require('promise-queue');
-const getS3EndpointForRegion = require('../deploy/lib/getS3EndpointForRegion');
+const getS3EndpointForRegion = require('../utils/getS3EndpointForRegion');
 
 const constants = {
   providerName: 'aws',
