+++ lib/plugins/aws/utils/getS3EndpointForRegion.test.js
